Modelapi/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
motor
certifi
dnspython
python-jose[cryptography]
python-multipart
passlib[bcrypt]
email-validator